requirements.txt
django~=4.1
djangorestframework~=3.13.1
psycopg2-binary
django-filter~=22.1
django-environ
#nginx==1.21.6
gunicorn~=20.1.0
pip~=22.0.4
pytz~=2022.2.1
asgiref~=3.5.2
sqlparse~=0.4.2
setuptools~=63.2.0
environ